Be smart re: github user name & prevent EEXISTS.
Fixes: https://github.com/bucharest-gold/entente/issues/47

The fix for this is maybe not ideal. It adds an additional key/value
pair to the package.json file: `gitHubUser: username`. But this allows
us to prompt for the name. I can't figure out how to actually use the
prompt (or get anything from `process.stdin`) outside the context of the
exported object. So, as far as I can see this is the best it will be. Of
course, I could be overlooking something. That happens.

Fixes: https://github.com/bucharest-gold/entente/issues/46

Now the whole thing doesn't blow up when a `test` directory already
exists. This change also prevents the process from overwriting any
existing file such as `LICENSE` or `index.js`.